% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbbarplot}
\alias{bbbarplot}
\title{Streamlines the creation of a barplot for primitive counting  using ggplot2}
\usage{
bbbarplot(data, factor, uniquecounts, countslab = "Anzahl", xlab = NULL,
  xrotate = FALSE, facet = NULL, prop = FALSE, xaxisorder = NULL,
  horizontal = FALSE, stackpar = NULL, facetncol = 2,
  stacktitle = "Gruppe", facetscales = "free", cex.datalabel = 2,
  datalabel = TRUE)
}
\arguments{
\item{data}{the dataframe to use}

\item{factor}{to group the data by}

\item{uniquecounts}{the variable to count}

\item{countslab}{lable of the y-axis}

\item{xlab}{label of x-axis}

\item{xrotate}{if true labels of x-axis are rotation by 90°}

\item{facet}{the factor to facet the plot by}

\item{prop}{if true the proportions will be ploted}

\item{xaxisorder}{vector of values to sort the values of "factor"}

\item{horizontal}{if ture a horizontal barplot is ploted}

\item{stackpar}{if the name of a variable is given a stacked barplot will be generated}

\item{facetncol}{the amount of column the facet plot is generated to}

\item{stacktitle}{the title of the stacked variable default="Gruppe"}

\item{facetscales}{the scale = "free"}

\item{cex.datalabel}{fontsize of datalabels default=2,}

\item{datalabel}{if true the data will be labeld default=TRUE}
}
\description{
Streamlines the creation of a barplot for primitive counting  using ggplot2
}
\examples{
PatID<-seq(1,13)
PatientWeightClass<-c("61-70","41-50","61-70","41-50","61-70","71-80","81-90","81-90","71-80","81-90","81-90","91-100","61-70")
PatientClass<-c("ClassA","ClassB","ClassB","ClassC","ClassA","ClassC", "ClassA","ClassB","ClassA","ClassC","ClassA","ClassA","ClassA")
df<-as.data.frame(cbind(PatID,PatientWeightClass,PatientClass))
p1<-bbbarplot(data=df,"PatientWeightClass","PatID",countslab="Anzahl", xlab=NULL,xrotate=FALSE,stackpar="PatientClass",stacktitle="PatientClass",datalabel=TRUE)
p1

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

