% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbChkDataFrame}
\alias{bbChkDataFrame}
\title{bbChkDataFrame}
\usage{
bbChkDataFrame(data)
}
\arguments{
\item{data}{a dataframe}
}
\description{
Checks weather a Dataframe is empty or not
}
\examples{
\dontrun{
df<-as.data.frame(c("a","b","c"))
bbChkDataFrame(df)
}

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
