% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbclearpage}
\alias{bbclearpage}
\title{Wrapperfuncction for latex '\\cleardoublepage' tag withhin markdown to control pagination in pdf output}
\usage{
bbclearpage()
}
\description{
Wrapperfuncction for latex '\\cleardoublepage' tag withhin markdown to control pagination in pdf output
}
\examples{
bbclearpage()

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
