% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbdedate}
\alias{bbdedate}
\title{bbdedate}
\usage{
bbdedate(x)
}
\arguments{
\item{x}{date string}
}
\description{
not done yet
}
\details{
Formats a Datestring to German Date format (dd-mm-yyyy)
}
\examples{
\dontrun{
bbdedate("2016-11-05")
}

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
