% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbDistinctCount}
\alias{bbDistinctCount}
\title{bbDistinctCount}
\usage{
bbDistinctCount(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Performs a distinct count of Items in a vector while omiting NAs. Basically a wrapper for dplyrs n_distinct
}
\details{
bbDistinctCount
}
\examples{
\dontrun{
bbDistinctCount(c(12,15,12,17,19,22,24,23,23,66,78,65,34,NA))
}

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
