% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbGetTpl}
\alias{bbGetTpl}
\title{bbGetTpl}
\usage{
bbGetTpl(templatefile)
}
\arguments{
\item{templatefile}{Name of Rmd or Rnw template}
}
\description{
Get the full Path for a Report Template based on the predefined Path by "bbSetTplPath"
}
\details{
Gibt unter Verwendung der Globaken Variablen TplPath das Reporttemplate zurueck
}
\examples{
\dontrun{
report<-bbGetTpl("c:/temp/Test.Rmd")
}

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
