% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbggboxplot}
\alias{bbggboxplot}
\title{boxplot with ggplot}
\usage{
bbggboxplot(data, factor, group, uniquecounts, ylabel = "Anzahl",
  xlabel = NULL, xrotate = FALSE, facet = NULL)
}
\arguments{
\item{data}{dataframe}

\item{factor}{the factor to count}

\item{group}{the group}

\item{uniquecounts}{the unique counts}

\item{ylabel}{the label of y axis default="Anzahl"}

\item{xlabel}{the label of x axis default=NULL}

\item{xrotate}{if TRUE the x and y axis will switch default=FALSE}

\item{facet}{if a facet is given plot will be faceted default=NULL}
}
\description{
boxplot with ggplot
}
\examples{
PatID<-seq(1,50)
PatientWeight<-runif(50, min=40, max=100)
PatientClass<-rep(c("ClassA","ClassB","ClassC","ClassB","ClassC"),10)
df<-as.data.frame(cbind(PatID,PatientWeight,PatientClass))
df$PatientWeight<-as.double(df$PatientWeight)
p1<-bbggboxplot(data=df,factor="PatientWeight",group="PatientClass",uniquecounts="PatID",ylabel="Anzahl",bin=1)
p1

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
