% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbgghistplot}
\alias{bbgghistplot}
\title{Streamlines plotting of a histogramm using ggplot2}
\usage{
bbgghistplot(data, factor, uniquecounts, countslab = "Anzahl",
  xlabel = NULL, bin = 1, facet = NULL, dens = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{factor}{the column to count}

\item{uniquecounts}{the column that determines the unique id}

\item{countslab}{label of y axis default="Anzahl"}

\item{xlabel}{label of x axis default=NULL}

\item{bin}{bandwidth default=1}

\item{facet}{if given the plot wil be printed as facets default=NULL}

\item{dens}{if TRUE a density plot wil be performed default=FALSE}
}
\description{
Streamlines plotting of a histogramm using ggplot2
}
\examples{
PatID<-seq(1,50)
PatientWeight<-runif(50, min=40, max=100)
PatientWeightClass<-c("61-70","41-50","61-70","41-50","61-70","71-80","81-90","81-90","71-80","81-90","81-90","91-100","61-70")
PatientClass<-c("ClassA","ClassB","ClassB","ClassC","ClassA","ClassC", "ClassA","ClassB","ClassA","ClassC","ClassA","ClassA","ClassA")
PatientClass2<-seq(1,50,50)
df<-as.data.frame(cbind(PatID,PatientWeight,PatientClass))
p1<-bbgghistplot(data=df,"PatientWeight","PatID",countslab="Anzahl",bin=10)
p1

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

