% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbreport}
\alias{bbreport}
\title{Wrapper for knit_expand}
\usage{
bbreport(file, ...)
}
\arguments{
\item{file}{pthe path to file}
}
\description{
Wrapper for knit_expand
}
\examples{
bbreport("c:/temp/Test.Rmd")

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
