% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbSetTplPath}
\alias{bbSetTplPath}
\title{bbSetTplPath}
\usage{
bbSetTplPath(x)
}
\arguments{
\item{x}{path path to folder}
}
\description{
not done yet
}
\details{
Set the Path to Report Template Folder as global variable
}
\examples{
\dontrun{
bbSetTplPath("c:/temp/Test")
}

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
