% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{ctab}
\alias{ctab}
\title{Creates a Crosstable with basic counts}
\usage{
ctab(row, col, margin = 1, dec = 1, percs = FALSE, total = FALSE)
}
\arguments{
\item{row}{vector containing factors for row}

\item{col}{vector containing factors for columns}

\item{margin}{index number (1 for rows, etc.)}

\item{dec}{decimal points (default =1)}

\item{total}{sum of columns and rows (default= FALSE)}
}
\description{
Creates a Crosstable with basic counts
}
\examples{
PatientCohort<-c("CohortA","CohortA","CohortA","CohortB","CohortB","CohortB","CohortB","CohortB","CohortB","CohortC","CohortC","CohortC","CohortC")
PatientClass<-c("ClassA","ClassB","ClassB","ClassC","ClassA","ClassC", "ClassA","ClassB","ClassA","ClassC","ClassA","ClassA","ClassA")
t<-ctab(row=PatientCohort,col=PatientClass,margin=1,total=T,perc=T)

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
