% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{oncosys_map}
\alias{oncosys_map}
\title{Plot of Map}
\usage{
oncosys_map(x, count.breaks = waiver(), shapefile, layer)
}
\arguments{
\item{x}{dataframew with long and lat}

\item{shapefile}{data source name (interpretation varies by driver — for some drivers, dsn is a file name, but may also be a folder)}

\item{layer}{layer name (varies by driver, may be a file name without extension)}
}
\description{
Plot of Map
}
\examples{
readOGR(dsn=paste(c(GetRDirPath(),"/Data/ShapeGER/"),collapse=""), layer="DEU_adm1",verbose=FALSE)


}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

