% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbclearpage}
\alias{bbclearpage}
\title{bbclearpage}
\usage{
bbclearpage()
}
\description{
not done yet
}
\details{
Wrapperfuncction for latex '\\cleardoublepage' tag withhin markdown to control pagination in pdf output
}
\examples{
\dontrun{
bbclearpage()
}

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
