% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbdedate}
\alias{bbdedate}
\title{Formats a Datestring to German Date format (dd-mm-yyyy)}
\usage{
bbdedate(x)
}
\arguments{
\item{x}{date string}
}
\description{
Formats a Datestring to German Date format (dd-mm-yyyy)
}
\examples{
bbdedate("2016-11-05")

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

