% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbDistinctCount}
\alias{bbDistinctCount}
\title{Performs a distinct count of Items in a vector (omits NA). Basically a wrapper for length() and unique()}
\usage{
bbDistinctCount(data)
}
\arguments{
\item{data}{a vector}
}
\description{
Performs a distinct count of Items in a vector (omits NA). Basically a wrapper for length() and unique()
}
\examples{
bbDistinctCount(c(12,15,12,17,19,22,24,23,23,66,78,65,34,NA))

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
