% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbGetTpl}
\alias{bbGetTpl}
\title{Get the full Path for a Report Template based on the predefined Path by "bbSetTplPath"}
\usage{
bbGetTpl(x)
}
\arguments{
\item{x}{Name of .Rmdd or .Rnw template}
}
\description{
Get the full Path for a Report Template based on the predefined Path by "bbSetTplPath"
}
\examples{
report<-bbGetTpl("c:/temp/Test.Rmd")

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
