% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbreport}
\alias{bbreport}
\title{bbreport}
\usage{
bbreport(file, ...)
}
\arguments{
\item{file}{pthe path to file}

\item{...}{other args}
}
\description{
not done yet
}
\details{
Wrapper for knit_expand
}
\examples{
\dontrun{
bbreport("c:/temp/Test.Rmd")
}

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
