% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportfunctions.R
\name{bbSetTplPath}
\alias{bbSetTplPath}
\title{Set the Path to Report Template Folder as global variable}
\usage{
bbSetTplPath(x)
}
\arguments{
\item{x}{path path to folder}
}
\description{
Set the Path to Report Template Folder as global variable
}
\examples{
bbSetTplPath("c:/temp/Test")

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}

